+++ docs/CHANGELOG.md
@@ -53,6 +53,7 @@ Essentially a rewrite of v1.*, adding/updating heaps of documentation and refact
     - Renamed some fields and methods within the `MoonPhase` data type.
 - Updated effect content structure.
 - Updated sound content structure.
+- Updated trinkets content structure.
 
 ### Content changes